% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryRaster.R
\name{wapor_queryRaster}
\alias{wapor_queryRaster}
\title{Query the FAO database for a specific raster product}
\usage{
wapor_queryRaster(
  collection = NULL,
  product = NULL,
  dimensions = list(),
  aoi = NULL,
  begin = NULL,
  end = NULL,
  APIkey = NULL,
  outdir = ".",
  cutline = FALSE,
  tiled = FALSE,
  compressed = FALSE,
  overviews = FALSE,
  sleep_time = 10
)
}
\arguments{
\item{collection}{A length one charachter vector indication the collection where
the queried product is found.}

\item{product}{A length one character vector indicating the desired product to download.}

\item{dimensions}{A list object specifying the dimensions of the product expect
of the time and measurment dimensions. Any other dimensions needs to be specified
in the form of \code{dimensions = list(KEY = "value")}. In cases of products
with only the temporal and measurement dimension an empty list should be handed
to the function. This is also the Default.}

\item{aoi}{An sf object which bounding box is used to clip the raster extent.
This is only relevant for regional or global datasets when clipping is desired.
Simply omit this option when you want to download a complete dataset (e.g.
a global dataset or a very local dataset such as WAPOR L3 data). The function
will issue a warning in these cases that the extent was set to global/regional.}

\item{begin}{A length one character in the form of \code{"YYYY-MM-DD"} or an
Date object in the same format. This time parameter is inclusive meaning that
all timestemps equal or higher than \code{begin} will be included.}

\item{end}{A length one character in the form of \code{"YYYY-MM-DD"} or an
Date object in the same format. This parameter is exclusive meaning that
all timestemps lower thatn \code{end} will be included.}

\item{APIkey}{A length one charachter containing your API key. Can be obtained
in the profile section at \href{https://wapor.apps.fao.org}{https://wapor.apps.fao.org}.}

\item{outdir}{A length one character vector with an existing directory. Any output
files will be written here.}

\item{cutline}{A logical indicating whether the resulting raster should be cut
to the bounding box of the \code{aoi} object. Only relevant when aoi is specified.}

\item{tiled}{Logical whether the resulting raster file should be tiled.}

\item{compressed}{Logical to indicate if the resulting raster should be compressed
using LZW.}

\item{overviews}{Logical to indicate if overviews for the resulting raster should
be created.}

\item{sleep_time}{A numeric value indicating the length of time R sleeps
when checking for the availability of the result in seconds. Depending
on the size of the queried raster (e.g. global extent at high spatial resolution)
it might be a good idea to increase the sleep time. Defaults to 10 seconds.}
}
\value{
nothing
}
\description{
This function is used to query a specific raster product in the FAO Gismanager.
Because data sets within the database vary greatly bettween the different
collection this function uses an easily abstractable pattern for the query.
This scheme follows the simple pattern of \strong{what} you want to download,
\strong{when} is the timeperiod you are interested in, and \strong{where} is the region
you want to download.
}
